refactor(parser): split text/structural document parsers into parser_pkg/text_blocks.py
Phase 1B Batch 8 (Move 19 from docs/PHASE1-PLAN.md).  Extract the
text and structural document parsers from parser.py into a new
mixin module.

Move 19 — parser_pkg/text_blocks.py
  Rules for: Paragraph (TEXT:), _parse_paragraph_coalesced,
  PureParagraph (PURETEXT:), LatexBlock (LATEX:),
  _parse_identifier_document_item, _parse_bracket_document_item,
  PartsDirective, BMachine (B:), RawLatexBlock, PageBreak,
  LineBreak, Contents, BibliographyMetadata, _parse_parts_format,
  Section, Solution, Part, TruthTable.

Notes on Moves 18 and 20:
* Move 18 (bindings) is absorbed: _parse_binding,
  _parse_binding_component, and _parse_theta already moved to
  parser_pkg/expressions.py in Batch 7 (Move 16), since the
  binding parser is part of the expression-grammar recursive
  descent.
* Move 20 (postfix) is absorbed: _parse_postfix already lives in
  parser_pkg/expressions.py from Batch 7 — it is the postfix layer
  of the recursive-descent expression grammar and stays with the
  rest of the grammar.

Gate 8 verified:
* make check — clean.
* make test-e2e — 159/159 byte-identical.
* make refactor-diff — 190/190 byte-identical vs main.

parser.py shrinks from 1467 → 948 lines (cumulative 85.1%).
