+++ src/txt2tex/parser_pkg/_base.py
@@ -42,6 +42,7 @@ class ParserError(Exception):
 
 if TYPE_CHECKING:
     from txt2tex.ast_nodes import (
+        Abbreviation,
         Aggregator,
         ArgueChain,
         AxDef,
@@ -59,6 +60,7 @@ if TYPE_CHECKING:
         GivenType,
         Group,
         GroupAggregate,
+        HorizDef,
         InfruleBlock,
         LatexBlock,
         LineBreak,
@@ -198,6 +200,7 @@ class ParserBase:
         def _parse_schema_rename_or_generic(
             self, base: Expr
         ) -> GenericInstantiation | SchemaRename: ...
+        def _parse_horiz_def(self) -> HorizDef: ...
         def _parse_horiz_def_generic_params(self) -> list[str] | None: ...
         def _parse_horiz_def_rhs(self) -> Expr | SchemaInclusion: ...
         def _parse_schema_body(
@@ -209,7 +212,7 @@ class ParserBase:
         def _parse_free_type(self) -> FreeType: ...
         def _parse_free_branch(self, branch_token: Token) -> FreeBranch: ...
         def _parse_syntax_block(self) -> SyntaxBlock: ...
-        def _parse_abbreviation(self) -> object: ...
+        def _parse_abbreviation(self) -> Abbreviation: ...
         def _parse_axdef(self) -> AxDef: ...
         def _parse_gendef(self) -> GenDef: ...
         def _parse_zed(self) -> Zed: ...
@@ -256,5 +259,5 @@ class ParserBase:
         def _parse_parts_format(self) -> str: ...
         def _parse_truth_table(self) -> TruthTable: ...
         def _parse_identifier_document_item(self) -> DocumentItem: ...
-        def _parse_bracket_document_item(self) -> DocumentItem: ...
+        def _parse_bracket_document_item(self) -> Expr | DocumentItem: ...
         def _parse_bibliography_metadata(self) -> BibliographyMetadata: ...
